created a login page which sends a POST request to the api/users/login (#16)
* created a login page which sends a POST request to the api/users/login
route. It performs password verification using argon2 and sets the
cookies. It will create a new token for the account if one does not
exist or it will use whatever one is already there for the account. This
token will persist through all the pages of the site until we log into
another account. I also changed the account registration to create a
token and give it to the site cookies. The hashing for the password has
been changed from being done in postgres queries to using argon2.

* ran prettier. these are the relevant affected files

* changed error from 400 to 409